2016/8/13+am 贺琳健订单接口预售和客服,现货

+++ app/src/main/java/com/silianchuangye/sumao/success/model/SpotOrderModel.java
@@ -9,7 +9,7 @@ public class SpotOrderModel implements Serializable {
 //    单价
     private double univalent;
 //    数量
-    private int number;
+    private double number;
 //    生产企业
     private String  enterprise;
 //    总价
@@ -29,11 +29,11 @@ public class SpotOrderModel implements Serializable {
         this.univalent = univalent;
     }
 
-    public int getNumber() {
+    public double getNumber() {
         return number;
     }
 
-    public void setNumber(int number) {
+    public void setNumber(double number) {
         this.number = number;
     }
